[RF69] Use 32-bit IRQ flags

+++ src/modules/RF69/RF69.cpp
@@ -259,7 +259,7 @@ int16_t RF69::startReceive() {
   return(state);
 }
 
-int16_t RF69::startReceive(uint32_t timeout, uint16_t irqFlags, uint16_t irqMask, size_t len) {
+int16_t RF69::startReceive(uint32_t timeout, uint32_t irqFlags, uint32_t irqMask, size_t len) {
   (void)timeout;
   (void)irqFlags;
   (void)irqMask;
